Rudimentry implementation of a ".graal" parser.
Nothing is hooked up to this, so no exporting to png just yet.  But its
close!

With a test level that doesn't have any repeating tiles as input, the
parser is able to fill out a full board without overflowing, and
correctly identifies the cut off between tile data and other level data.

Levels feature repeating tiles, however, crash.

There is enough here that I could try to hook it up to the image
converter stuff later and see if the output is anywhere near correct.
This might also engender refactoring the nw parser a bit and providing a
common base class between the two.

This relates to issue #12.