Refactored project to be more object-oriented.
There is now a common level class, and an Actor class.  This commit also
moves a lot of scarping functionality off of nw_parser.py into the
common.py module, so that it can be used by other formats.

The two current parsers are not interchangable just yet, but this is
really close to being able to parse the tile data from some .graal
formatted levels.

This, unfortunately introduces a bug, where seteffect images (or maybe
just all npcs) don't show up in the final render.

This relates to issue #5, and to issue #12.