DotGraalParser populates board now (buggy).
The DotGraalParser now populates its board.  So, now some levels are
renderable with the now inaccurately named nw2png.py script.

This is, however, but close.  The output is wrong, because it assumes
the tile encoding would have something in common with the .nw format,
but copying the four lines for that didn't have the desired effect.

This relates to issue #12.