Fixed tile parsing of .graal files.
It seems likely to be correct, though I've not verified the actual
output for correctness.  But all of my test levels now parse, and the
generated packets look sensible.

My mistake was in the assumption of how bits in the file map to packets.
Instead, the bytes in the file are left-shifted on to the end of a
buffer, and then the buffer is right-shifted to pull packets off of it.
It looks like the file format was desgined so that the tiles could be
decoded concurrently with IO.

Anyways, it works now.  Mine approach is definitely not suitable for
streaming, because get_packet works without creating side effects, but
the internals are kind of similar.