Add support for .graal and .zelda format revisions
This commit adds support for all of the common revisions of the
.graal/.zelda file format.  The parser is still only limited to tile
data, though, but the packet size is detected automatically now based on
the file header.

This commit also refactors the common paths for level loading, so that
parsing level data is separate from detecting valid file headers.  This
DRYs things up a bit, but is mainly so that parse-time assertion
failures are not interpreted as a negative file header match.