Corrected .graal parsing rules, almost complete.
Links, baddies, and NPCs are probably correct now, or only need minor
adjustment.

Hopefully this should also account for all of the differences between
versions, as well as handling garbage data correctly.  It seems that
1.01 and 1.02 at least had some problems with this.

Treasure boxes are now supported, too, and partial support exists for
signs now.

So what is missing is essentially just:

 - sign text encoding

 - testing