refactor analysis tool for better development iteration
This change splits setlist.py so that you call it once to generate a
level database (which is slow), and then a second time to generate a
level playlist (which is fast).

This mainly is so that its easier to tinker with the playlist
generator.

This change also makes the overall behavior deterministic, which is
nice.

Main outstanding issue right now is that the quad matching code will
cause Problems if a level path is a hash alias.  I think a better way
to handle this is to not do all of my lookups by hash, and include
"duplicate" levels in the by_path dict.  This would effectively
exclude them from being matched randomly, but make them possible to
hit otherwise.

I think the use for the "boring" list should only be to prevent the
level from being linked to, and possibly exclude it from by_hash, but
otherwise leave it in by_path.

One last area for improvement is making the debug output for what is
"skipped" more helpful.  Eg, note what levels have been merged
together and to what level they were merged into.